+++ NEWS.md
@@ -3,7 +3,9 @@
 * `simpol1()` gains an argument `meredith` that controls whether just the functional groups in the original SIMPOL.1 method (Pankow & Asher, 2008) is used or if additional coefficients used in Meredith et al. (2023) are also included.
 * The default for the `method` argument to `calc_vol()` has now been renamed to `"meredith"`.  `"simpol1"` now uses the original SIMPOL.1 method without additional coefficients added in Meredith et al. (2023).
 * It is now possible to supply input to `calc_vol()` as a vector of SMILES strings with `from = "smiles"`
+* Users can now choose from RVI thresholds for non-volatile, low, moderate, and high volatility for clean atmosphere, polluted atmosphere, or soil using the `environment` parameter of `calc_vol()
 * `simpol1()` now takes into account amide functional groups.
+
 
 # volcalc 2.0.0